# openapi.yaml
openapi: 3.1.0
info:
  title: geo
  version: "1.0.0"
  description: >-
    IP geolocation, reference data (countries, currencies, time zones) and phone/distance helpers,
    plus your own place collections with nearby search. IP geolocation reads MaxMind-format (MMDB)
    files configured via MMDB_PATH/ASN_MMDB_PATH; when MMDB_PATH is unset, IP lookups still work
    for special (private/loopback/link-local/multicast/reserved/unspecified) addresses but return
    503 for public ones. A freshly opened MMDB file is checksum-verified (against MMDB_SHA256 /
    ASN_MMDB_SHA256, or a `<path>.sha256` sidecar in standard sha256sum format, when either is
    configured) and then validated with real tree-walk lookups before it ever replaces the reader
    currently serving traffic — both at process start and on a manual reload
    (POST /v1/database/reload). A failure at any of those stages leaves the previously loaded
    database, if any, untouched and serving.
security:
  - ApiKey: []
paths:
...
  /v1/ip/self:
    get:
      operationId: geo.ip.lookupSelf
      summary: Look up the caller's own IP address
      description: >-
        Looks up `request.ip` (the client address Fastify resolves, honoring TRUST_PROXY). Same
        lookup logic and response shape as `GET /v1/ip/{ip}`.
      tags: [ip]
      x-required-role: read
      parameters:
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          $ref: '#/components/responses/IpLookupOk'
        '400':
          $ref: '#/components/responses/InvalidIp'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /v1/ip/{ip}:
    get:
      operationId: geo.ip.lookup
      summary: Look up an IP address
      description: >-
        Resolves country, region, city, coordinates, time zone and (when ASN_MMDB_PATH is
        configured) ASN/organization for a public IPv4 or IPv6 address. Special-range addresses
        (private, loopback, link-local, multicast, reserved, unspecified — `kind` in the response)
        are classified without consulting the MMDB file and always return `found: false` with every
        geo field `null`, never a 503, even when no IP database is configured at all. A public
        address returns 503 DATABASE_UNAVAILABLE only when no city database is configured or the
        currently loaded one failed; the ASN lookup, when configured, augments the result but never
        causes a 503 on its own (a missing/failed ASN database just leaves `asn: null`).
      tags: [ip]
      x-required-role: read
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 64
          example: 1.1.1.1
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          $ref: '#/components/responses/IpLookupOk'
        '400':
          $ref: '#/components/responses/InvalidIp'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /v1/ip/batch:
    post:
      operationId: geo.ip.lookupBatch
      summary: Look up multiple IP addresses
      description: >-
        Two-layer size cap: the JSON schema allows up to 1000 items (`maxItems: 1000`) as an outer
        ceiling — exceeding it fails the whole request with 400 VALIDATION_FAILED before the
        handler runs. The real operative cap is the `MAX_BATCH` environment variable (default 100,
        configurable 1-1000); a request between `MAX_BATCH` and 1000 items passes schema validation
        but is then rejected whole with 413 BATCH_TOO_LARGE by the handler.

        **Per-item errors do not fail the batch.** For each IP that passes both caps, any domain
        error the single-IP lookup would have thrown (`INVALID_IP`, or — most importantly —
        `DATABASE_UNAVAILABLE`) is caught per item and reported as `{ ip, ok: false, error: {
        code, message } }` inside an overall HTTP 200. This is a real asymmetry with
        `GET /v1/ip/{ip}` / `GET /v1/ip/self`: there, `DATABASE_UNAVAILABLE` is a genuine HTTP 503;
        here, the identical condition never changes the batch's HTTP status — it only ever appears
        as a per-item `ok: false` entry.
      tags: [ip]
      x-required-role: read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IpBatchRequest'
      responses:
        '200':
          description: >-
            One result per input IP, in the same order, each independently ok or failed (see
            description above — a per-item failure never changes this 200).
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpBatchResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/BatchTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/database:
    get:
      operationId: geo.database.getInfo
      summary: Current IP database status
      description: Whether a city/ASN database is loaded, its metadata, when it was loaded, and lookup counters since process start.
      tags: [database]
      x-required-role: read
      responses:
        '200':
          $ref: '#/components/responses/DatabaseInfoOk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/database/reload:
    post:
      operationId: geo.database.reload
      summary: Reload the configured IP database file(s) from disk
      description: >-
        Re-reads MMDB_PATH/ASN_MMDB_PATH. Each candidate file is checksum-verified (when
        MMDB_SHA256/ASN_MMDB_SHA256 or a `<path>.sha256` sidecar is configured), opened, and
        validated with real lookups for known public IPs (1.1.1.1, 8.8.8.8) — a thrown error at any
        stage rejects that candidate. Both candidates are fully validated into local variables
        before either replaces the reader currently serving, so a failure here never disturbs
        already-loaded databases: on failure the previous readers (if any) keep serving unchanged
        and this returns 409 DATABASE_RELOAD_FAILED. On success, the response is the same shape as
        `GET /v1/database`, and the next `GET /ready` re-checks immediately instead of using its
        10-second cache.
      tags: [database]
      x-required-role: write
      responses:
        '200':
          $ref: '#/components/responses/DatabaseInfoOk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/DatabaseReloadFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/countries:
    get:
      operationId: geo.countries.list
      summary: List countries
      description: Optionally filtered by name/code prefix, continent, EU membership or currency.
      tags: [countries]
      x-required-role: read
      parameters:
        - $ref: '#/components/parameters/Lang'
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
          description: Case/accent-folded prefix match against the localized name, alpha-2 or alpha-3 code.
        - name: continent
          in: query
          schema:
            type: string
            enum: [AF, AN, AS, EU, NA, OC, SA, af, an, as, eu, na, oc, sa]
        - name: eu
          in: query
          schema: { type: string, enum: ['true', 'false'] }
        - name: currency
          in: query
          schema: { type: string, pattern: '^[A-Za-z]{3}$' }
      responses:
        '200':
          description: Countries matching the filter, sorted by localized name.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CountryView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/countries/{code}:
    get:
      operationId: geo.countries.get
      summary: Get a country by code
      description: Looked up by alpha-2, alpha-3 or numeric code, case-insensitive.
      tags: [countries]
      x-required-role: read
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, minLength: 2, maxLength: 3 }
          example: TR
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          description: The matched country.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [country]
                properties:
                  country:
                    $ref: '#/components/schemas/CountryView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CountryNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/currencies:
    get:
      operationId: geo.currencies.list
      summary: List currencies
      description: Every currency used by at least one country in the reference table, sorted by ISO code.
      tags: [currencies]
      x-required-role: read
      parameters:
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          description: All currencies.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CurrencyView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/currencies/{code}:
    get:
      operationId: geo.currencies.get
      summary: Get a currency by ISO 4217 code
      description: 404s unless the code is used by at least one country in the reference table (not every ISO 4217 code is accepted).
      tags: [currencies]
      x-required-role: read
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, minLength: 2, maxLength: 3 }
          example: TRY
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          description: The matched currency.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [currency]
                properties:
                  currency:
                    $ref: '#/components/schemas/CurrencyView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CurrencyNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/timezones:
    get:
      operationId: geo.timezones.list
      summary: List time zones
      description: Every IANA zone name Node's ICU knows plus every zone in the reference table (their union), optionally filtered.
      tags: [timezones]
      x-required-role: read
      parameters:
        - $ref: '#/components/parameters/Lang'
        - name: country
          in: query
          schema: { type: string, minLength: 2, maxLength: 3 }
          description: Restrict to the zones of one country (alpha-2/alpha-3/numeric). 404s (COUNTRY_NOT_FOUND) if the country is unknown.
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
          description: Case-insensitive substring match against the zone name.
      responses:
        '200':
          description: Matching time zones, evaluated at the current instant.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimezoneView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CountryNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/phone:
    get:
      operationId: geo.phone.normalize
      summary: Normalize a phone number to E.164
      description: >-
        National numbering plans are not modelled: validity only checks that the resolved calling
        code is known and the digit count is 8-15 (E.164 length rule), not that the subscriber
        number actually exists. When `number` has no `+`/`00` international prefix, `country` is
        required to resolve it as a national number; if `country` is given but not a known country
        code, this returns 404 COUNTRY_NOT_FOUND (not a 400) — the same code
        `GET /v1/countries/{code}` uses. An unresolvable international calling code, or a missing
        `country` for a national-format number, is not an error: it returns 200 with `valid: false`
        and a `reason`.
      tags: [phone]
      x-required-role: read
      parameters:
        - name: number
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 40 }
          example: "+90 532 000 00 00"
        - name: country
          in: query
          schema: { type: string, minLength: 2, maxLength: 3 }
          description: Default country for a number given without an international prefix.
          example: TR
      responses:
        '200':
          $ref: '#/components/responses/PhoneNormalizeOk'
        '400':
          $ref: '#/components/responses/InvalidPhone'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CountryNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/phone/batch:
    post:
      operationId: geo.phone.normalizeBatch
      summary: Normalize multiple phone numbers
      description: >-
        Same two-layer size cap as `POST /v1/ip/batch`: JSON schema allows up to 1000 items
        (`maxItems: 1000`) as an outer ceiling (exceeding it is 400 VALIDATION_FAILED before the
        handler runs); the real operative cap is `MAX_BATCH` (default 100) — between `MAX_BATCH`
        and 1000 items is rejected whole with 413 BATCH_TOO_LARGE.

        Per-item errors do not fail the batch: any domain error a single normalize call would throw
        for one number — `INVALID_PHONE`, or `COUNTRY_NOT_FOUND` when the shared `country` is not a
        known code — is caught per item and reported as `{ number, ok: false, error: { code,
        message } }` inside an overall HTTP 200, exactly like `POST /v1/ip/batch`. Phone
        normalization has no database dependency, so unlike the IP batch route there is no
        `DATABASE_UNAVAILABLE` case here.
      tags: [phone]
      x-required-role: read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneBatchRequest'
      responses:
        '200':
          description: One result per input number, in the same order, each independently ok or failed.
          headers:
            X-Content-Type-Options:
              $ref: '#/components/headers/XContentTypeOptions'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneBatchResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/BatchTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <secret>`. Keys come from `GEO_API_KEYS`
        (`id:secret[:role]`, role one of `read`/`write`/`readwrite`, default `readwrite`), compared
        in constant time across every configured key (no early exit). Every operation below
        declares its required role via the `x-required-role` extension (`read` or `write`;
        `readwrite` always satisfies either). Required on every `/v1/*` route and on `/metrics`;
        `/health`, `/ready` and `/v1/info` are unauthenticated.
  parameters:
    Lang:
      name: lang
      in: query
      schema:
        type: string
        minLength: 2
        maxLength: 35
        pattern: '^[A-Za-z0-9-]+$'
      description: >-
        BCP 47 language tag for localized names. Invalid or unrecognized tags never error — they
        silently fall back to DEFAULT_LANG (env-configurable, default "en").
...
  responses:
    IpLookupOk:
      description: "The lookup result. `found: false` for a special-range address or a public address not covered by the database."
      headers:
        X-Content-Type-Options:
          $ref: '#/components/headers/XContentTypeOptions'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IpLookupResult'
...
  headers:
    XContentTypeOptions:
      description: Set on every response by geo's onSend hook.
      schema:
        type: string
        const: nosniff
...